perf: bound semantic model inference
