+++ docs/reference/public-surface.md
@@ -202,6 +202,9 @@ The configuration inventory follows `.env.sample`; unlisted implementation-only 
 - SCRAPE_CACHE_VOLATILE_CAP
 - SEARXNG_URL
 - SEMANTIC_URL
+- SEMANTIC_INFERENCE_ADMISSION_TIMEOUT_SECONDS
+- SEMANTIC_INFERENCE_QUEUE_SIZE
+- SEMANTIC_INFERENCE_WORKERS
 - SECTION_FILTER_DEFAULT_EXCLUDE
 - SECTION_FILTER_DEFAULT_INCLUDE
 - SECTION_FILTER_DEFAULT_VERBOSITY
