+++ semantic-svc/app.py
@@ -34,15 +34,19 @@ import logging
 import math
 import os
 import time
-from collections.abc import Callable
 from contextlib import asynccontextmanager
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
+import inference as inference_module
 import numpy as np
 from auth import verify_api_key
 from fastapi import Depends, FastAPI, HTTPException, Request, Response
 from fastapi.responses import JSONResponse
+from inference import (
+    InferenceManager,
+    InferenceOverloaded,
+    run_inference,
+)
 from metrics import METRICS
 from models import (
     EmbedRequest,
@@ -61,6 +65,7 @@ from common.middleware import add_request_id_middleware
 
 setup_logging()
 logger = logging.getLogger(__name__)
+_inference_manager = inference_module._inference_manager
 
 
 def _load_sentence_transformers():
@@ -116,305 +121,6 @@ class TaskTracker:
             await asyncio.gather(*pending, return_exceptions=True)
 
 
-# ── Bounded native inference execution ────────────────────────────
-
-
-def _positive_int_env(name: str, default: int, *, minimum: int = 1) -> int:
-    """Read a bounded integer setting and fail fast on invalid deployment config."""
-    value = int(os.getenv(name, str(default)))
-    if value < minimum:
-        raise ValueError(f"{name} must be >= {minimum}")
-    return value
-
-
-def _nonnegative_int_env(name: str, default: int) -> int:
-    """Read a queue-size setting that may be zero (no waiting queue)."""
-    value = int(os.getenv(name, str(default)))
-    if value < 0:
-        raise ValueError(f"{name} must be >= 0")
-    return value
-
-
-INFERENCE_WORKERS = _positive_int_env("SEMANTIC_INFERENCE_WORKERS", 1)
-INFERENCE_QUEUE_SIZE = _nonnegative_int_env("SEMANTIC_INFERENCE_QUEUE_SIZE", 16)
-INFERENCE_ADMISSION_TIMEOUT = float(
-    os.getenv("SEMANTIC_INFERENCE_ADMISSION_TIMEOUT_SECONDS", "0.25")
-)
-if INFERENCE_ADMISSION_TIMEOUT <= 0:
-    raise ValueError("SEMANTIC_INFERENCE_ADMISSION_TIMEOUT_SECONDS must be > 0")
-
-_INFERENCE_OPERATIONS = frozenset(
-    {
-        "rerank",
-        "embed",
-        "vector_search",
-        "index",
-        "index_batch",
-        "migration_backfill",
-    }
-)
-_INFERENCE_PRIORITIES = {"interactive": 0, "maintenance": 10}
-
-
-class InferenceOverloadedError(Exception):
-    """Raised when bounded inference admission cannot accept a request."""
-
-    def __init__(self, operation: str, retry_after: float) -> None:
-        self.operation = operation
-        self.retry_after = retry_after
-        super().__init__(f"semantic inference capacity exhausted for {operation}")
-
-
-# Short compatibility name for callers that want to catch admission overloads.
-InferenceOverloaded = InferenceOverloadedError
-
-
-@dataclass
-class _InferenceWork:
-    """One admitted native call and its result future."""
-
-    operation: str
-    priority: int
-    function: Callable[[], Any]
-    result: asyncio.Future[Any]
-    admitted_at: float
-    abandoned: bool = False
-    native_started: bool = False
-    sequence: int = field(default=0)
-
-
-class InferenceManager:
-    """Run model calls in a bounded, priority-aware native worker pool.
-
-    The semaphore bounds both active calls and queued calls. Native futures are
-    awaited by manager-owned workers, so cancellation of an HTTP task never
-    releases a capacity slot while a blocking model call is still running.
-    """
-
-    def __init__(
-        self,
-        *,
-        max_workers: int = INFERENCE_WORKERS,
-        queue_size: int = INFERENCE_QUEUE_SIZE,
-        admission_timeout: float = INFERENCE_ADMISSION_TIMEOUT,
-    ) -> None:
-        if max_workers < 1:
-            raise ValueError("max_workers must be >= 1")
-        if queue_size < 0:
-            raise ValueError("queue_size must be >= 0")
-        if admission_timeout <= 0:
-            raise ValueError("admission_timeout must be > 0")
-        self.max_workers = max_workers
-        self.queue_size = queue_size
-        self.admission_timeout = admission_timeout
-        self._capacity = asyncio.Semaphore(max_workers + queue_size)
-        self._queue: asyncio.PriorityQueue[tuple[int, int, _InferenceWork | None]] = (
-            asyncio.PriorityQueue()
-        )
-        self._workers: list[asyncio.Task[None]] = []
-        self._loop: asyncio.AbstractEventLoop | None = None
-        self._sequence = 0
-        self._closed = False
-        self._queued = 0
-        self._in_flight = 0
-
-    @staticmethod
-    def _operation_name(operation: str) -> str:
-        return operation if operation in _INFERENCE_OPERATIONS else "other"
-
-    def _observe_queue_depth(self) -> None:
-        METRICS.gauge(
-            "groktocrawl_semantic_inference_queue_depth",
-            "Number of admitted model calls waiting for a native worker",
-        ).set(value=float(self._queued))
-
-    def _observe_in_flight(self) -> None:
-        METRICS.gauge(
-            "groktocrawl_semantic_inference_in_flight",
-            "Number of native model calls currently running",
-        ).set(value=float(self._in_flight))
-
-    async def _ensure_started(self) -> None:
-        if self._closed:
-            raise InferenceOverloadedError("other", self.admission_timeout)
-        loop = asyncio.get_running_loop()
-        if self._loop is not None and self._loop is not loop:
-            # Direct unit callers may use asyncio.run repeatedly. Production
-            # lifespan uses one loop; reset only an idle manager between loops.
-            if self._in_flight or self._queued:
-                raise InferenceOverloadedError("other", self.admission_timeout)
-            self._workers = []
-            self._capacity = asyncio.Semaphore(self.max_workers + self.queue_size)
-            self._queue = asyncio.PriorityQueue()
-        self._loop = loop
-        if not self._workers:
-            self._workers = [
-                asyncio.create_task(self._worker()) for _ in range(self.max_workers)
-            ]
-
-    async def run(
-        self,
-        operation: str,
-        function: Callable[[], Any],
-        *,
-        priority: str = "interactive",
-    ) -> Any:
-        """Admit and execute one blocking model call.
-
-        ``asyncio.shield`` deliberately prevents cancellation from propagating
-        to the manager-owned result future. The native call therefore remains
-        accounted for until it actually returns or raises.
-        """
-        operation = self._operation_name(operation)
-        priority_value = _INFERENCE_PRIORITIES.get(priority, 0)
-        await self._ensure_started()
-        admission_start = time.monotonic()
-        try:
-            await asyncio.wait_for(
-                self._capacity.acquire(), timeout=self.admission_timeout
-            )
-        except TimeoutError as exc:
-            METRICS.counter(
-                "groktocrawl_semantic_inference_overloads_total",
-                "Model calls rejected after bounded admission timeout",
-                ["operation"],
-            ).inc({"operation": operation})
-            raise InferenceOverloadedError(operation, self.admission_timeout) from exc
-
-        if self._closed:
-            self._capacity.release()
-            raise InferenceOverloadedError(operation, self.admission_timeout)
-
-        loop = asyncio.get_running_loop()
-        item = _InferenceWork(
-            operation=operation,
-            priority=priority_value,
-            function=function,
-            result=loop.create_future(),
-            admitted_at=admission_start,
-            sequence=self._sequence,
-        )
-        self._sequence += 1
-        self._queued += 1
-        self._observe_queue_depth()
-        self._queue.put_nowait((item.priority, item.sequence, item))
-        try:
-            return await asyncio.shield(item.result)
-        except asyncio.CancelledError:
-            # The manager worker continues to own the slot and native future.
-            item.abandoned = True
-            METRICS.counter(
-                "groktocrawl_semantic_inference_cancellations_total",
-                "HTTP callers canceled while model inference was admitted",
-                ["operation"],
-            ).inc({"operation": operation})
-            raise
-
-    async def _worker(self) -> None:
-        while True:
-            if self._closed and self._queue.empty():
-                return
-            _, _, item = await self._queue.get()
-            if item is None:
-                self._queue.task_done()
-                return
-            self._queued -= 1
-            self._observe_queue_depth()
-            if item.abandoned:
-                self._capacity.release()
-                self._queue.task_done()
-                continue
-
-            queue_wait = time.monotonic() - item.admitted_at
-            METRICS.histogram(
-                "groktocrawl_semantic_inference_queue_wait_seconds",
-                "Time spent admitted before native model execution",
-                ["operation", "priority"],
-            ).observe(
-                {
-                    "operation": item.operation,
-                    "priority": "maintenance" if item.priority else "interactive",
-                },
-                queue_wait,
-            )
-            item.native_started = True
-            self._in_flight += 1
-            self._observe_in_flight()
-            native_start = time.monotonic()
-            try:
-                # The manager owns exactly ``max_workers`` worker tasks. The
-                # loop's executor is used so asyncio.run can drain it during
-                # test/process shutdown; admission still bounds native calls.
-                loop = asyncio.get_running_loop()
-                result = await loop.run_in_executor(None, item.function)
-            except BaseException as exc:
-                if not item.abandoned and not item.result.done():
-                    item.result.set_exception(exc)
-                if not isinstance(exc, asyncio.CancelledError):
-                    METRICS.counter(
-                        "groktocrawl_semantic_inference_failures_total",
-                        "Native model calls that raised an exception",
-                        ["operation"],
-                    ).inc({"operation": item.operation})
-            else:
-                if not item.abandoned and not item.result.done():
-                    item.result.set_result(result)
-            finally:
-                METRICS.histogram(
-                    "groktocrawl_semantic_inference_duration_seconds",
-                    "Native model inference latency",
-                    ["operation"],
-                ).observe(
-                    {"operation": item.operation}, time.monotonic() - native_start
-                )
-                self._in_flight -= 1
-                self._observe_in_flight()
-                self._capacity.release()
-                self._queue.task_done()
-
-    async def shutdown(self) -> None:
-        """Reject queued work, drain native calls, then close the executor."""
-        if self._closed:
-            return
-        self._closed = True
-        while not self._queue.empty():
-            _, _, item = self._queue.get_nowait()
-            if item is None:
-                self._queue.task_done()
-                continue
-            self._queued -= 1
-            item.abandoned = True
-            if not item.result.done():
-                item.result.cancel()
-            self._capacity.release()
-            self._queue.task_done()
-        self._observe_queue_depth()
-        if self._workers:
-            # Wake idle workers. Active workers finish their native call first,
-            # then consume one sentinel each and exit.
-            for _ in self._workers:
-                self._queue.put_nowait((100, self._sequence, None))
-                self._sequence += 1
-            await asyncio.gather(*self._workers, return_exceptions=True)
-
-
-_inference_manager = InferenceManager()
-
-
-def _get_inference_manager() -> InferenceManager:
-    return _inference_manager
-
-
-async def run_inference(
-    operation: str,
-    function: Callable[[], Any],
-    *,
-    priority: str = "interactive",
-) -> Any:
-    """Execute a model call through the process-wide bounded inference pool."""
-    return await _get_inference_manager().run(operation, function, priority=priority)
-
-
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     """Load SentenceTransformer and CrossEncoder models at startup."""
@@ -423,6 +129,7 @@ async def lifespan(app: FastAPI):
     # A fresh manager is tied to this event loop. Its workers drain native
     # calls before the executor is closed during shutdown.
     _inference_manager = InferenceManager()
+    inference_module._inference_manager = _inference_manager
     app.state.inference_manager = _inference_manager
     logger.info("Loading semantic models (~2.2GB, 2-5s)...")
     loop = asyncio.get_event_loop()
